Refactor code to meet Java coding standard
