+++ data/Karen.txt
@@ -1,2 +1,3 @@
 T | 0 | bake bread
 D | 1 | poop | 2023-10-11 1200
+T | 1 | go sleep
