+++ src/main/java/karen/commands/Command.java
@@ -1,7 +1,7 @@
-package Karen.commands;
+package karen.commands;
 
-import Karen.tasks.TaskList;
-import Karen.util.Ui;
+import karen.tasks.TaskList;
+import karen.util.Ui;
 
 public abstract class Command {
     public abstract void execute(TaskList taskList, Ui ui);
