+++ src/main/java/karen/commands/DeleteCommand.java
@@ -1,22 +1,22 @@
-package Karen.commands;
+package karen.commands;
 
-import Karen.tasks.Task;
-import Karen.tasks.TaskList;
-import Karen.util.Storage;
-import Karen.util.Ui;
+import karen.tasks.Task;
+import karen.tasks.TaskList;
+import karen.util.Storage;
+import karen.util.Ui;
 
-public class UnmarkCommand extends Command{
+public class DeleteCommand extends Command {
     private int index;
 
-    public UnmarkCommand(int i) {
+    public DeleteCommand(int i) {
         this.index = i;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui) {
-        taskList.unmarkTask(this.index);
         Task t = taskList.getTask(this.index);
-        ui.showUnmarkMessage(t);
+        taskList.deleteTask(this.index);
+        ui.showDeleteMessage(t);
         Storage.saveToFile(taskList);
     }
