+++ src/main/java/karen/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Karen.tasks;
+package karen.tasks;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
