+++ src/test/java/karen/tasks/TaskListTest.java
@@ -1,4 +1,4 @@
-package Karen.tasks;
+package karen.tasks;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertArrayEquals;
