+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 _______________________
-Hi! I'm Karen
+Hi! I'm karen
 What can I do for you?
 _______________________
 _______________________
